[EI-191] [web] refactor: Refactor searchedIndicatorList in source-indicator-search-list
